Cleanup test files
The get_free_port function wasn't used anywhere else and the
environment is setup is done differently now.
